requirements.txt
Flask
google-genai
google-generativeai
numpy
pypdf
pysqlite3
python-dotenv
pytest